+++ frontend/src/serverspec-gen/ast-interface.ts
@@ -1,0 +1,28 @@
+/// <reference path="../../declares.d.ts" />
+
+export interface Describe {
+  resourceType: string;
+  name: string;
+  body: (It|Its)[];
+}
+
+export interface It {
+  type: string;
+  should: boolean;
+  matcher: Matcher;
+}
+
+export interface Its extends It {
+  name: string;
+}
+
+export interface Matcher {
+  name: string;
+  args: string[];
+  chains: Chain[];
+}
+
+export interface Chain {
+  name: string;
+  args: string[];
+}
