+++ frontend/tsconfig.json
@@ -13,6 +13,7 @@
     "src/helper.ts",
     "src/server_status.ts",
     "src/users_admin.ts",
+    "src/serverspec-gen/ast-interface.ts",
     "src/serverspec-gen/ast.ts",
     "src/serverspec-gen/it-panel.ts",
     "src/serverspec-gen/resource-panel.ts",
